# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  description: 轻量级任务追踪系统 RESTful API
  version: 1.0.0
  contact:
    name: Task Tracker
servers:
  - url: http://localhost:3300
    description: 本地开发

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string

    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string

    Tag:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        color:
          type: string
        task_count:
          type: integer

    Task:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        due_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'

    TaskCreateInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
          default: todo
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        due_date:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: integer

    TaskUpdateInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        due_date:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: integer

    BatchInput:
      type: object
      required: [action, ids]
      properties:
        action:
          type: string
          enum: [update_status, update_priority, delete]
        ids:
          type: array
          items:
            type: integer
          minItems: 1
        value:
          type: string
          description: 新的状态或优先级值（action为delete时不需要）

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  /api/health:
    get:
      summary: 健康检查
      tags: [System]
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /api/auth/register:
    post:
      summary: 用户注册
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: admin
                email:
                  type: string
                  format: email
                  example: admin@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 用户名或邮箱已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      summary: 用户登录
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks:
    get:
      summary: 获取任务列表
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: "按状态筛选，多个用逗号分隔 (todo,in_progress,done)"
        - name: priority
          in: query
          schema:
            type: string
          description: "按优先级筛选，多个用逗号分隔"
        - name: tag
          in: query
          schema:
            type: string
          description: "按标签ID筛选，多个用逗号分隔"
        - name: search
          in: query
          schema:
            type: string
          description: 搜索标题和描述
        - name: due_before
          in: query
          schema:
            type: string
            format: date
          description: 截止日期早于
        - name: due_after
          in: query
          schema:
            type: string
            format: date
          description: 截止日期晚于
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, due_date, priority, title, status]
            default: created_at
          description: 排序字段
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: 排序方向
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: 每页条数
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: 创建任务
      tags: [Tasks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string

  /api/tasks/{id}:
    get:
      summary: 获取单个任务
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: 更新任务
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: 删除任务
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks/batch:
    post:
      summary: 批量操作
      tags: [Tasks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInput'
            examples:
              batch_complete:
                summary: 批量标记完成
                value:
                  action: update_status
                  ids: [1, 2, 3]
                  value: done
              batch_delete:
                summary: 批量删除
                value:
                  action: delete
                  ids: [4, 5]
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  affected:
                    type: integer
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string

  /api/tags:
    get:
      summary: 获取标签列表
      tags: [Tags]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 标签列表（含任务计数）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      summary: 创建标签
      tags: [Tags]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 50
                color:
                  type: string
                  pattern: '^#[0-9a-fA-F]{6}$'
                  default: '#6366f1'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '409':
          description: 标签名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tags/{id}:
    put:
      summary: 更新标签
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: 标签不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: 删除标签
      tags: [Tags]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: 标签不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'